requirements.txt
blinker==1.6.2
cachelib==0.10.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.3.2
Flask-Login==0.6.2
Flask-Session==0.5.0
Flask-WTF==1.1.1
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
Misskey.py==4.1.0
nanoatp==0.3.4
oauthlib==3.2.2
Pillow==9.5.0
requests==2.30.0
requests-oauthlib==1.3.1
tld==0.13
tweepy==4.14.0
urllib3==2.0.2
Werkzeug==2.3.4
WTForms==3.0.1
